Changed EOF caclulation and Datatools bugfix
The scipy.sparse.linalg function was producing inconsistent EOFs even
thought the singular values were the same.  I'm guessing because my
data was by no means sparse and the iterative nature of the method are
at fault, though I am by no means sure.  I switched to scipy.linalg.svd
instead which is slower.  It's a one-off calculation in the LMR though,
so it really doesn't change things that much.  The DataTools bugfix
changed where the dimensions of the field are loaded from.  Instead
of the Dataset, I load them from the variable, for which, they should
already be in order.