Removed DataObject Usage From LIM
The LIM was doing too much work that should be controlled by the
user.  Replaced the explicit interaction with BaseDataObject-type
to just use ndarrays.  This means the user will now have to make
sure that data is processed consistently for use with a LIM.

Added in some basic logging to the runtime of the LIM and its
forecast.