Copy and preliminary test-train split function
Implemented a copy function to produce a replica of the DataObject
with the (potentially subsetted) current databin.  For HDF5 DataObjects
this will create a new default group for the copy in the same file.

The test train split function preliminary code has been  added. Uses
the copy feature to provide training datasets, but testing sets are
just the arrays.