Add seed parameter to noise_integration
When running noise integration using multiprocessing (uses a fork
operation) the numpy RNG is all initialized the same, so I get the
same result for every process.  In order to prevent this, noise
integration now accepts a seed which should be different for each
time the method is called if you're using multiprocessing.